[REFACTOR/#364] Kakao Util 추가하여 리팩토링
