+++ feature/intern/build.gradle.kts
@@ -11,10 +11,10 @@ android {
 dependencies {
     // domain
     implementation(projects.domain.intern)
+    implementation(libs.kakao.share)
 
     // feature
     implementation(projects.feature.dialog)
-    implementation(libs.kakao.share)
 
     testImplementation(projects.core.testing)
 }
